split functions for presenting instructions and exiting out into separate scripts
